requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
python-dateutil==2.9.0.post0
